Work In Progress
Added to .gitignore the logs folder and the geckodriver log.
Added tqdm dependency to the requirements file.
Created the class GalleryItem.py representing a gallery item.

Added in the options file the PATH to the webdriver.
Added in the options file the headers, parameters, CDN and XPATH paths
to download and get the items.

Configured in FirefoxConfigurator the method to start the webdriver
correctly.

Added in the DownloadManager the method to scroll the gallery and the
method to get the information of each item.

Abstracted in NetworkManager the method to make requests and created the
method to download the items.

Created in StorageManager the method to write to disk the items and
updated the method to create the download folder.

Simplified in main.py the method process_arguments by creating methods
process_galleries and execute_downloader.